The BioChemSolenoidValve is now controlled by any Relay component, not limited to the SwitchBoxMPIKG device.
This design enables consistent control of multiple devices connected to any electronic box that exposes a Relay component. The devices attached to the relay are not locked to the MPIKG box!

Updated the documentation to clarify how users should define support_platform as device_name/component_name and specify channel when applicable.